Fixing POSE_LANDMARK_COUNT bitfield size
n D:\Github\git-BrendanWalker\MikanMediaPipe (branch body_pose), BodyPoseObservation::providedMask in src/Vision/BodyPoseTypes.h is a uint32_t but POSE_LANDMARK_COUNT is 33. Two shift expressions are undefined behavior:

1. src/App/TrackingJson.cpp:194 — outBody.providedMask= j.value("providedMask", (1u << POSE_LANDMARK_COUNT) - 1u); MSVC emits warning C4293 for this and computes 1u << 33 as 1u << 1, so the "all slots provided" default for pre-mask recordings becomes 1, i.e. only landmark 0 reads as provided. Any recording made before providedMask was added replays with an almost-empty skeleton.
2. src/Vision/BodyPoseTypes.h:89 — isProvided() does 1u << landmarkIndex, so index 32 (RIGHT_FOOT_INDEX) can never be reported provided. Nothing currently fills slot 32 (RTMPose maps into indices up to 28), so this one is latent.

Fix: widen providedMask to uint64_t and use 1ull shifts throughout (check src/App/TrackingJson.cpp, src/App/DiagnosticDump.cpp, src/UI/HandOverlay.cpp, src/Vision/BodyPoseTracker.cpp and the tests for other places that build or read the mask). Confirm the JSON round-trip still works for a 33-bit value (nlohmann handles uint64 fine, but check the recorded field's read path). Rebuild, then run --selftest, --test-bodypose and --test-replay, and --replay-verify on a recording in %APPDATA%/MikanMediaPipe/recordings to confirm nothing diverges. Note that CMake uses file(GLOB), so re-run cmake -S . -B build if you add files.

+++ src/Vision/BodyPoseTypes.h
@@ -83,10 +83,13 @@ struct BodyPoseObservation
 	// different topology (RTMPose emits COCO's 17) leaves the rest untouched
 	// at the image origin - and an overlay that drew them would pin a phantom
 	// landmark in the corner of the frame.
-	uint32_t providedMask= 0;
+	//
+	// 64-bit because there are 33 slots: a 32-bit mask cannot address the
+	// last one, and shifting by 33 is undefined rather than merely wrong.
+	uint64_t providedMask= 0;
 	bool isProvided(int landmarkIndex) const
 	{
-		return (providedMask & (1u << landmarkIndex)) != 0;
+		return (providedMask & (1ull << landmarkIndex)) != 0;
 	}
 
 	// sigmoid [0,1]: within frame and not occluded
